+++ frontend/src/components/Project/ProjectDetail/ProjectDetail.test.tsx
@@ -1,7 +1,7 @@
 import { fireEvent, render, screen } from "@testing-library/react";
 import { Provider } from "react-redux";
 import ProjectDetail from "./ProjectDetail";
-import { store } from "../../store";
+import { store } from "../../../store";
 import userEvent from "@testing-library/user-event";
 
 describe("ProjectDetail", () => {
