+++ frontend/src/components/Scrollbar/styles.ts
