+++ frontend/src/components/Styles/GlobalStyles.tsx
