+++ frontend/src/components/Table/CollapsibleTable.tsx
