+++ frontend/src/containers/Home/Home.tsx
@@ -5,11 +5,11 @@ import GridLayout from "./GridView/ProjectGridView";
 import NotiPieChart from "./PieChart/PieChart";
 import RecentThree from "./RecentThree/RecentThree";
 import Upcoming from "./Upcoming/Upcoming";
-import Scrollbar from "../../components/scrollbar/Scrollbar";
+import Scrollbar from "../../components/Scrollbar/Scrollbar";
 import { authSelector } from "../../store/slices/auth";
 import { fetchProjects } from "../../store/slices/project";
 import { AppDispatch } from "../../store";
-import ProjectCreateModal from "../../components/project/ProjectCreateModal";
+import ProjectCreateModal from "../../components/Project/ProjectCreateModal";
 import List from "./ListView/ProjectListView";
 import Button from "@mui/material/Button";
 import Stack from "@mui/material/Stack";
